Added Map Functions to Facilitate Tile-Based Aiming (Untested)

+++ sprite.h
@@ -80,7 +80,7 @@ class Sprite
 		// whether sprite is trying to aim in-hand item
 		bool aiming = false;
 		// coordinates being aimed at. Only valid when aiming = true
-		float aimingX = 0, aimingY = 0;
+		int aimingX = 0, aimingY = 0;
 		
 		// item sprite is holding in-hand. Sprite is drawn to reflect it is holding the item
 		Item* inHand = NULL;
